LPS-138126 Move StringPlus logic to StringUtil
